♻️ KC9ANF task: decompose release plan helpers
