+++ .agentplane/tasks/202605290621-KC9ANF/pr/diffstat.txt
@@ -1,0 +1,3 @@
+ .../src/commands/release/plan.command.ts           | 122 ++-------------------
+ .../src/commands/release/plan.helpers.ts           | 113 +++++++++++++++++++
+ 2 files changed, 124 insertions(+), 111 deletions(-)
