+++ .agentplane/tasks/202605290621-KC9ANF/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 KC9ANF task: Release plan command decomposition [202605290621-KC9ANF]
