+++ .agentplane/tasks/202605290621-KC9ANF/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605290621-KC9ANF/release-plan-command-decomposition",
+  "created_at": "2026-05-29T06:21:37.555Z",
+  "diffstat_sha256": "sha256:f648ec6d3a3a9c0a26ce083b910b1411b063ba1da3688a1cbc0f5062845ad684",
+  "last_verified_at": "2026-05-29T06:26:03.144Z",
+  "last_verified_diffstat_sha256": "sha256:e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855",
+  "schema_version": 1,
+  "task_id": "202605290621-KC9ANF",
+  "updated_at": "2026-05-29T06:21:37.555Z",
+  "verify": {
+    "status": "pass"
+  }
+}
