+++ .agentplane/tasks/202605290621-KC9ANF/pr/review.md
@@ -1,0 +1,38 @@
+# PR Review
+
+Created: 2026-05-29T06:21:37.555Z
+
+## Task
+
+- Task: `202605290621-KC9ANF`
+- Title: Release plan command decomposition
+- Status: DOING
+- Branch: `task/202605290621-KC9ANF/release-plan-command-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290621-KC9ANF/README.md`
+
+## Verification
+
+- State: ok
+- Note: Release plan helpers extracted into packages/agentplane/src/commands/release/plan.helpers.ts; runReleasePlan behavior and generated release plan artifacts are preserved. Checks passed: release plan/apply matrix (31 pass); bun run typecheck; bun run arch:check; bun run knip:check; bun run lint:core; bun run format:changed; bun run hotspots:check (runtime hotspots 3 -> 2).
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T06:21:37.555Z
+- Branch: task/202605290621-KC9ANF/release-plan-command-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+ .../src/commands/release/plan.command.ts           | 122 ++-------------------
+ .../src/commands/release/plan.helpers.ts           | 113 +++++++++++++++++++
+ 2 files changed, 124 insertions(+), 111 deletions(-)
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
